feat(vfs): extend VfsStat — mtime/nlink/uid/gid/blocks; batched readdir
VfsStat carries full POSIX-ish metadata (was {size: usize, mode: usize}).
New fields: mtime/nlink/uid/gid/blocks. readdir returns (name, stat)
pairs in one round trip — eliminates N+1 stat for ls -l, du, find.

Wire format v2: 40-byte stat payload for stat/fstat replies; per-entry
readdir layout [name_len:u32 LE][stat:40 bytes][name] replacing the old
[name_len:u8][is_dir:u8][name].

ext2 backend reads full inode metadata (mtime/nlink/uid/gid/blocks).
RemoteBackend does N+1 FS_STAT calls to enrich each readdir entry.
ramfs/memfs/procfs/devfs synthesize sensible defaults. FileStat trait
extended to carry the new fields; virtio-blk FS_STAT reply extended to
6 words packing mtime/nlink/uid/gid.

Phase 4 Plan C Stage 1.

+++ userspace/libcluu/src/posix/file.rs
@@ -517,9 +517,9 @@ fn vfs_fstat(entry: &mut FdEntry) -> Option<usize> {
     };
     match vfs_client.fstat(vfs_file) {
         Ok(stat) => {
-            entry.file_size = Some(stat.size);
-            entry.file_mode = Some(stat.mode);
-            Some(stat.size)
+            entry.file_size = Some(stat.size as usize);
+            entry.file_mode = Some(stat.mode as usize);
+            Some(stat.size as usize)
         }
         Err(_) => None,
     }
